Add the '-a' option to 'zpool export'.
Support exporting all imported pools in one go, using 'zpool export -a'.

This is accomplished by moving the export parts from zpool_do_export()
in to the new function zpool_export_one().  The for_each_pool() function
is used to enumerate the list of pools to be exported.  Passing an argc
of 0 implys the function should be called on all pools.

Signed-off-by: Turbo Fredriksson <turbo@bayour.com>
Closes: #3203