Update all default taskq settings
Over the years the default values for the taskqs used on Linux have
differed slightly from illumos.  In the vast majority of cases this
was done to avoid creating an obnoxious number of idle threads which
would polute the process listing.

With the addition of support for dynamic taskqs all multi-threaded
queues should be created as dynamic taskqs.  This allows us to get
the best of both worlds.

* The illumos default values can be restored which is known to
improve the performance of certain workloads.

* Reduce the number of idle threads on the system when it's not
under heavy load.  The maximum number of threads will only be
created when they are required.

This change also increases the maximum number of threads in the
asynchronous iput taskq.  This has been observed to be a bottleneck
for certain workloads.  For consistency with the other zfs pipeline
threads 'zfs_iput_taskq' threads have been renamed 'z_iput'.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>