Add zfs_sb_prune_compat() function
For kernels which do not implement a per-suberblock shrinker the
shrink_dcache_parent() function was used to attempt to reclaim
dentries.  This was found not be entirely reliable which could
lead to performance issues on older kernels running meta-data
heavy workloads.

To address this issue a zfs_sb_prune_compat() function has been
added to implement this functionality.  It relies on traversing
the list of inodes for a filesystem and calling d_prune_alias()
to drop all dentries.  A call to zfs_iput_async() is used to drop
the last reference on the inode which allows it to be cleaned up
safely outside the z_znodes_lock.  To provide the same synchronous
behavior of the per-filesystem shrinker the function blocks waiting
for the iputs to compelte.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>