Allow rollback to arbitrary txg
This patch will disable the logic which prevents you from rolling
back to an arbitrary txg.  Normally a txg which is older than the
last label update cannot be used for import.  There is no way to
know that the given txg is consistent with the current configuration.
This should only be used as a last resort for pool recovery when
you encounter the following failure mode.

$ zpool import -o readonly=on -T 123456 tank
cannot import 'tank': one or more devices is currently unavailable

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #2452