Dump unique configurations and Uberblocks in zdb -lu
For zdb -l, detect when the configuration nvlist in some label l (l>0)
is the same as a configuration already dumped.  If so, simply print
"Configuration same as label X" where X is the identical config printed
earlier, instead of dumping the configuration.

Make a similar check when dumping Uberblocks for zdb -lu.  For a given
slot S, check whether a label already dumped contains an identical
Uberblock in the same slot S.  If so, do not dump the Uberblock.

An example of the output:
	------------------------------------
	LABEL 0
	------------------------------------
	    version: 5000
	    name: 'testpool.22001'
	    state: 0
	    txg: 4
	    pool_guid: 10359279877037374982
	    errata: 0
	    hostname: 'locks'
	    top_guid: 2603524987218334357
	    guid: 8308788017679220434
	    vdev_children: 1
	    vdev_tree:
		type: 'mirror'
		id: 0
		guid: 2603524987218334357
		metaslab_array: 69
	    ...
	    <additional output redacted>
	    ...
	    Uberblock[8]
		magic = 0000000000bab10c
		version = 5000
		txg = 8
		guid_sum = 6258178503038716282
		timestamp = 1486593615 UTC = Wed Feb  8 14:40:15 2017
	------------------------------------
	LABEL 1
	------------------------------------
	    Configuration same as label 0
	    Uberblock[4] same as label 0
	    Uberblock[5] same as label 0
	    Uberblock[6] same as label 0
	    Uberblock[8] same as label 0
	    Uberblock[9]
		magic = 0000000000bab10c
		version = 5000
		txg = 9
		guid_sum = 6258178503038716282
		timestamp = 1486593615 UTC = Wed Feb  8 14:40:15 2017
	    ...
	    <additional output redacted>

Add two tests, zdb_00{3,4}_pos to verify the above functionality.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>

Store config and uberblock checksums in avl_tree_t

Store checksums for uberblocks and configurations in an avl tree,
and record which labels they appear in.  Use this to determine the
sets of unique configurations and unique uberblocks.

Instead of making a single pass over each label, make two passes.

1) Read the labels into an array of buffers.  Calculate and
record the checksums of each Uberblock and each buffer holding a
packed configuration nvlist.

2) Iterate over the labels and dump the labels and uberblocks
according to the options given by the user, as follows:

-l      Dump each unique configuration only once.
        Indicate which labels it appears in.
-ll     In addition, dump label space usage stats.
-ll     Dump every configuration, unique or not.

-u      Dump each unique, valid, uberblock only once.
        Indicate which labels it appears in.
-uu     In addition, state which slots are invalid.
-uuu    Dump every uberblock, unique or not.
-uuuu   Dump the uberblock blockpointer (used to be -uuu)