Fix l2arc_evict() destroy race
When destroying an arc_buf_hdr_t its identify cannot be discarded
until it is entirely undiscoverable.  This not only includes being
unhashed, but also being removed from the l2arc header list.
Discarding the header's identify prematurely renders the hash
lock useless because it will always hash to bucket zero.

This change resolves a race with l2arc_evict() by discarding the
identity after it has been removed from the l2arc header list.
This ensures either the header is not on the list or contains
the correct identify.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>