Linux 5.0 compat: SIMD compatibility (WIP)
Restore SIMD optimizations for 5.0 and newer kernels.  This is
accomplished by leveraging the fact that by definition dedicated
kernel threads will never need to save/restore the FPU state of
user space.  Only preemption and local interrupts need to be
disabled to allow the FPU to be used.

This is not a major restriction for ZFS since the vast majority
of SIMD operations are already performed by the IO pipeline.
The remaining couple of cases are relatively infrequent and can
be handled by the generic implementations without any significant
impact.  The two most noteworthy cases are:

  1) Decrypting the wrapping key for an encrypted dataset,
     i.e. `zfs load-key`.  All other encryption and decryption
     operations will use the SIMD optimized implementations.

  2) Generating the payload checksums for a `zfs send` stream.

In order to avoid making any changes to the higher layers of ZFS
all of the `*_get_ops()` functions were updated to take in to
consideration the calling context.  This allows for the fastest
implementation to be used as appropriate (see kfpu_allowed()).

The only other notable instance of SIMD operations being used
outside a kernel thread was at module load time.  This code
was moved in to a taskq in order to accomidate the new kernel
thread restriction.

Finally, a few other modifications were made in order to further
harden this code and facilite testing.  They include updating
each implementations operations structure to be declared as a
constant.  And allowing "cycle" to be set when selecting the
preferred ops in the kernel as well as user space.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>