DEBUG: rwlock reader tracking
Introduce a new RW_READER_PIDS lock type for debugging.  Locks
created with this type behave as normal but additionally allocate
a bitmap indexed by process pid to track all active readers.
When bit N is set it indicates the pid N is holding the lock as
a reader, otherwise it is clear.  By default, the bitmap is
sized for 32,768 pids which inflates the rwlock size by 4k.
It may be increased as needed when debugging.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>