simd: use alloc_pages_node to force alignment
fxsave and xsave require the target address to be 16-/64-byte aligned.

kmalloc(_node) does not (yet) offer such fine-grained control over
alignement[0,1], even though it does "the right thing" most of the time
for power-of-2 sizes. unfortunately, alignment is completely off when
using certain debugging or hardening features/configs, such as KASAN,
slub_debug=Z or the not-yet-upstream SLAB_CANARY.

Use alloc_pages_node() instead which allows us to allocate page-aligned
memory.  Since pages are a minimum of 4K this is guaranteed to satisfy
the alignment constraints.

Fixes: #9608

0: https://lwn.net/Articles/787740/
1: https://lore.kernel.org/linux-block/20190826111627.7505-1-vbabka@suse.cz/

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>Foo
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>